+++ mono_repo/lib/src/commands/github/action_info.dart
@@ -1,3 +1,5 @@
+import '../../root_config.dart';
+import 'action_versions.dart';
 import 'job.dart';
 import 'step.dart';
 
@@ -5,29 +7,29 @@ enum ActionInfo implements Comparable<ActionInfo> {
   cache(
     name: 'Cache Pub hosted dependencies',
     repo: 'actions/cache',
-    version: '88522ab9f39a2ea568f7027eddc7d8d8bc9d59c8', // v3.3.1
+    version: actionsCacheVersion,
   ),
   checkout(
     name: 'Checkout repository',
     repo: 'actions/checkout',
-    version: '8e5e7e5ab8b370d6c329ec480221332ada57f0ab', // v3.5.2
+    version: actionsCheckoutVersion,
   ),
   setupDart(
     name: 'Setup Dart SDK',
     repo: 'dart-lang/setup-dart',
-    version: 'd6a63dab3335f427404425de0fbfed4686d93c4f', // v1.5.0
+    version: dartLangSetupDartVersion,
   ),
   setupFlutter(
     name: 'Setup Flutter SDK',
     repo: 'subosito/flutter-action',
-    version: '48cafc24713cca54bbe03cdc3a423187d413aafa', // v2.10.0
+    version: subositoFlutterActionVersion,
   ),
 
   /// See https://github.com/marketplace/actions/coveralls-github-action
   coveralls(
     name: 'Upload coverage to Coveralls',
     repo: 'coverallsapp/github-action',
-    version: 'master',
+    version: coverallsappGithubActionVersion,
     completionJobFactory: _coverageCompletionJob,
   ),
 
@@ -35,7 +37,7 @@ enum ActionInfo implements Comparable<ActionInfo> {
   codecov(
     name: 'Upload coverage to codecov.io',
     repo: 'codecov/codecov-action',
-    version: 'main',
+    version: codecovCodecovActionVersion,
   );
 
   const ActionInfo({
@@ -48,16 +50,19 @@ enum ActionInfo implements Comparable<ActionInfo> {
   final String repo;
   final String version;
   final String name;
-  final Job Function()? completionJobFactory;
+  final Job Function(RootConfig rootConfig)? completionJobFactory;
 
   Step usage({
     String? name,
     String? id,
     Map<String, dynamic>? withContent,
+    Map<String, String>? versionOverrides,
   }) {
     name ??= this.name;
+    final useVersion =
+        (versionOverrides == null ? null : versionOverrides[repo]) ?? version;
     final step = Step.uses(
-      uses: '$repo@$version',
+      uses: '$repo@$useVersion',
       id: id,
       name: name,
       withContent: withContent,
@@ -71,7 +76,7 @@ enum ActionInfo implements Comparable<ActionInfo> {
   int compareTo(ActionInfo other) => index.compareTo(other.index);
 }
 
-Job _coverageCompletionJob() => Job(
+Job _coverageCompletionJob(RootConfig rootConfig) => Job(
       name: 'Mark Coveralls job finished',
       runsOn: 'ubuntu-latest',
       steps: [
@@ -82,6 +87,7 @@ Job _coverageCompletionJob() => Job(
             'github-token': r'${{ secrets.GITHUB_TOKEN }}',
             'parallel-finished': true
           },
+          versionOverrides: rootConfig.existingActionVersions,
         )
       ],
     );