+++ mono_repo/lib/src/commands/github/github_yaml.dart
@@ -122,7 +122,7 @@ Map<String, String> generateGitHubYml(RootConfig rootConfig) {
         Map.fromEntries(allJobs.map((e) => MapEntry(e.id, e.value)));
 
     for (var completion in completionMap.entries) {
-      final job = completion.key.completionJobFactory!()
+      final job = completion.key.completionJobFactory!(rootConfig)
         ..needs = completion.value.toList();
 
       jobList['job_${jobList.length + 1}'] = job;
@@ -204,7 +204,10 @@ Iterable<_MapEntryWithStage> _listJobs(
 
   for (var job in jobs) {
     if (job is _SelfValidateJob) {
-      yield jobEntry(_selfValidateJob(rootConfig.monoConfig), job.stageName);
+      yield jobEntry(
+        _selfValidateJob(rootConfig.monoConfig, rootConfig),
+        job.stageName,
+      );
       continue;
     }
 
@@ -340,6 +343,7 @@ extension on CIJobEntry {
       job.flavor,
       job.sdk,
       commandEntries,
+      rootConfig,
       config: rootConfig.monoConfig,
       additionalCacheKeys: {
         'packages': packages.join('-'),
@@ -382,7 +386,8 @@ Job _githubJob(
   String runsOn,
   PackageFlavor packageFlavor,
   String sdkVersion,
-  List<_CommandEntryBase> runCommands, {
+  List<_CommandEntryBase> runCommands,
+  RootConfig rootConfig, {
   required BasicConfiguration config,
   Map<String, String>? additionalCacheKeys,
 }) =>
@@ -393,17 +398,20 @@ Job _githubJob(
         if (!runsOn.startsWith('windows'))
           _cacheEntries(
             runsOn,
+            rootConfig: rootConfig,
             additionalCacheKeys: {
               'sdk': sdkVersion,
               if (additionalCacheKeys != null) ...additionalCacheKeys,
             },
           ),
-        packageFlavor.setupStep(sdkVersion),
+        packageFlavor.setupStep(sdkVersion, rootConfig),
         ..._beforeSteps(runCommands.whereType<_CommandEntry>()),
         ActionInfo.checkout.usage(
           id: 'checkout',
+          versionOverrides: rootConfig.existingActionVersions,
         ),
-        for (var command in runCommands) ...command.runContent(config),
+        for (var command in runCommands)
+          ...command.runContent(config, rootConfig),
       ],
     );
 
@@ -424,7 +432,7 @@ class _CommandEntryBase {
 
   _CommandEntryBase(this.name, this.run);
 
-  Iterable<Step> runContent(BasicConfiguration config) =>
+  Iterable<Step> runContent(BasicConfiguration config, RootConfig rootConfig) =>
       [Step.run(name: name, run: run)];
 }
 
@@ -444,7 +452,8 @@ class _CommandEntry extends _CommandEntryBase {
   });
 
   @override
-  Iterable<Step> runContent(BasicConfiguration config) => [
+  Iterable<Step> runContent(BasicConfiguration config, RootConfig rootConfig) =>
+      [
         Step.run(
           id: id,
           name: name,
@@ -452,7 +461,7 @@ class _CommandEntry extends _CommandEntryBase {
           workingDirectory: workingDirectory,
           run: run,
         ),
-        ...?type?.afterEachSteps(workingDirectory, config),
+        ...?type?.afterEachSteps(workingDirectory, config, rootConfig),
       ];
 }
 
@@ -464,6 +473,7 @@ class _CommandEntry extends _CommandEntryBase {
 /// store and retrieve the cache.
 Step _cacheEntries(
   String runsOn, {
+  required RootConfig rootConfig,
   Map<String, String>? additionalCacheKeys,
 }) {
   final cacheKeyParts = [
@@ -490,6 +500,7 @@ Step _cacheEntries(
       'key': restoreKeys.first,
       'restore-keys': restoreKeys.skip(1).join('\n'),
     },
+    versionOverrides: rootConfig.existingActionVersions,
   );
 }
 
@@ -500,7 +511,8 @@ String _maxLength(String input) {
   return input.substring(0, 512 - hash.length) + hash;
 }
 
-Job _selfValidateJob(BasicConfiguration config) => _githubJob(
+Job _selfValidateJob(BasicConfiguration config, RootConfig rootConfig) =>
+    _githubJob(
       selfValidateJobName,
       _ubuntuLatest,
       PackageFlavor.dart,
@@ -509,6 +521,7 @@ Job _selfValidateJob(BasicConfiguration config) => _githubJob(
         for (var command in selfValidateCommands)
           _CommandEntryBase(selfValidateJobName, command),
       ],
+      rootConfig,
       config: config,
     );
 
@@ -537,16 +550,18 @@ class _MapEntryWithStage {
 }
 
 extension on PackageFlavor {
-  Step setupStep(String sdkVersion) {
+  Step setupStep(String sdkVersion, RootConfig rootConfig) {
     switch (this) {
       case PackageFlavor.dart:
         return ActionInfo.setupDart.usage(
           withContent: {'sdk': sdkVersion},
+          versionOverrides: rootConfig.existingActionVersions,
         );
 
       case PackageFlavor.flutter:
         return ActionInfo.setupFlutter.usage(
           withContent: {'channel': sdkVersion},
+          versionOverrides: rootConfig.existingActionVersions,
         );
     }
   }